backend/requirements.txt
fastapi
uvicorn
python-multipart
numpy
becquerel
scipy
matplotlib
pyserial
websockets
reportlab

